add the styles
