+++ backend/src/api/controllers/userController.js
@@ -1,5 +1,5 @@
 import User from "../models/userModel.js";
-import logger from "../../utils/logger.js";
+
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
 const SECRET_KEY = "RECIPEAPI";
@@ -54,12 +54,12 @@ const signin = async (req, res) => {
     const token = jwt.sign(
       { email: existingUser.email, id: existingUser._id },
       SECRET_KEY,
-      { expiresIn: "1m" }
+      { expiresIn: "1h" }
     );
 
     res.status(201).json({ result: existingUser, token });
     /* return user data */
-    res.status(200).json({ result });
+    // res.status(200).json({ result });
   } catch (error) {
     console.log(error);
     return res.status(500).json({ message: "Something went wrong" });
