+++ backend/src/api/routes/userRoutes.js
@@ -1,5 +1,5 @@
 import express from "express";
-import logger from "../../utils/logger.js";
+
 import { signup, signin } from "../controllers/userController.js";
 
 const userRouter = express.Router();
