+++ frontend/src/App.css
@@ -8,7 +8,7 @@
 }
 
 body {
-  background-image: linear-gradient(to right, #0f0c29, #2d2767, #24243e);
+  background-image: linear-gradient(to right, #0f0c29, #241f4f, #24243e);
 }
 
 @media (prefers-reduced-motion: no-preference) {
