+++ frontend/src/styles/Footer.css
@@ -3,7 +3,7 @@
   border: 0.1px solid #008e7b8a;
   padding: 20px;
   background-image: linear-gradient(to right, #16222a, #3a6073);
-  margin-top: -50px;
+  margin-top: 50px;
   margin-bottom: 1px;
   color: aliceblue;
   font-family: monospace;
