+++ frontend/src/styles/LoginScreen.css
@@ -48,9 +48,11 @@ span {
   color: #ffffff;
   font-size: 12px;
   font-weight: bold;
-  padding: 12px 120px;
+  padding: 12px 50px;
   letter-spacing: 1px;
   text-transform: uppercase;
+  margin-top: 0px;
+  margin-left: 25%;
   transition: transform 60ms ease-in;
 }
 
@@ -65,7 +67,7 @@ span {
 .login-input-boxe {
   background-color: #323232;
   border: 0.5px solid black;
-  padding: 12px 40px;
+  padding: 10px 40px;
   margin: 8px 0;
   letter-spacing: 1px;
