+++ frontend/src/styles/texttospeesh.css
@@ -4,7 +4,7 @@
   padding-bottom: 277px;
   margin-top: 50px;
   text-align: center;
-  border: 0.1px solid #008e7b;
+
   background-image: linear-gradient(to right, #0f0c29, #241f4f, #24243e);
 }
